requirements.txt
setuptools==69.5.1
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.111.0
orjson==3.10.3
pydantic==2.7.1
backoff==2.2.1
uvicorn==0.29.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
python-dotenv==1.0.1
gunicorn==22.0.0
